Get first heading from current file
In order to quickly debug while prototyping. A command was pinned via
`Command Palette` and was used while iterating through test code until
it was possible to resolve the `heading[0]` from a `TFile` object.

I relied heavily on an 'obsidian api class diagram' [^1] and a code
snippet `app.metadataCache.getFileCache(file)?.links` [^2] to figure
this out.

[^1]: https://i.joethei.space/obsidian-api.png
[^2]: https://forum.obsidian.md/t/how-to-get-backlinks-for-a-file/45314