hotfix for setTabsTitle not updating in edge case
had an issue while testing where it would render headings fine but if i removed just the heading and nothing else. It wouldn't update the displayed title. I debugged every function related to this until I finally figured out that the event `app.vault.on('modify')` can't detect that scenario but `app.metadataCache.on('changed')` can. So, I'm playing with the two.